+++ infrastructure/terraform/per_workspace/modules/api_entrypoint/api_gateway/debug.tf
@@ -1,4 +1,4 @@
 resource "local_file" "rendered_swagger" {
-  content  = sensitive(data.template_file.swagger.rendered)
+  content  = sensitive(local.swagger_file)
   filename = "${path.root}/../../swagger/dist/aws/rendered/swagger.yaml"
 }